Enforce compliant formatting in CI
With recent changes we have formatted the code base using rustfmt and
included the corresponding configuration in the repository. With this
change we take things one step further and check formatting in CI. Note
that we have to use rustfmt from the nightly channel, just because the
stable version supports way too few useful configuration options.